Convert project to PlatformIO and fix some formatting errors

# Conflicts:
#	.gitignore
#	Motor.h
#	lib/Motor/Motor.cpp
#	src/ROS-motor-control.cpp

+++ lib/Motor/Motor.cpp
@@ -1,8 +1,8 @@
 #include "Motor.h"
 
-Motor::Motor(uint8_t IN1, uint8_t IN2, uint8_t PWM, uint8_t STBY, uint8_t encoderA, uint8_t encoderB, bool reverse, double kp, double ki, double kd)
-  : IN1(IN1), IN2(IN2), PWM(PWM), STBY(STBY), A(encoderA), B(encoderB), reverse(reverse)
-{
+Motor::Motor(uint8_t IN1, uint8_t IN2, uint8_t PWM, uint8_t STBY,
+  uint8_t encoderA, uint8_t encoderB, bool reverse, double kp, double ki, double kd)
+  : IN1(IN1), IN2(IN2), PWM(PWM), STBY(STBY), A(encoderA), B(encoderB), reverse(reverse) {
   /*pid = new PID(&input, &output, &setpoint,kp,ki,kd, DIRECT);
   pid->SetSampleTime(SAMPLE_TIME);
   pid->SetOutputLimits(-OUTPUT_LIMIT, OUTPUT_LIMIT);
@@ -21,33 +21,28 @@ void Motor::init() {
   //setPidEnabled(true);
 }
 
-// Enable PID control or pass 
-void Motor::setPidEnabled(bool enable)
-{
+// Enable PID control or pass
+void Motor::setPidEnabled(bool enable) {
   pid->SetMode(enable ? AUTOMATIC : MANUAL);
 }
 
-void Motor::setTunings(double kp, double ki, double kd)
-{
+void Motor::setTunings(double kp, double ki, double kd) {
   pid->SetTunings(kp, ki, kd);
 }
 
-long Motor::getPosition()
-{
+int32_t Motor::getPosition() {
   return encoder->read();
 }
 
 // Update the target velocity
-void Motor::setVelocity(double vel)
-{
+void Motor::setVelocity(double vel) {
   setpoint = vel;
 }
 
-void Motor::update()
-{
-  long currentPos = encoder->read();
-  long currentTime = millis();
-  long dT = currentTime - lastTime;
+void Motor::update() {
+  int32_t currentPos = encoder->read();
+  int32_t currentTime = millis();
+  int32_t dT = currentTime - lastTime;
   input = (double)(currentPos - lastPos)/(double)dT;
   lastPos = currentPos;
   lastTime = currentTime;
@@ -65,8 +60,7 @@ void Motor::update()
 }
 
 // Set motor power
-void Motor::write(double power)
-{
+void Motor::write(double power) {
   // Todo: check all this. Maybe add something to handle both sides better
   // Both high when vel == 0 which makes it brake.
   // Both low would be free spin which is kinda the same on gear motors
